Generate default avatar hue.

+++ Codebase/clucker-server/src/main/java/com/clucker/cluckerserver/model/User.java
@@ -61,7 +61,7 @@ public class User {
     private String avatarImage;
 
     @Max(360)
-    private int avatarHue;
+    private double avatarHue;
 
     @CreationTimestamp
     private LocalDateTime joined;
